android can start

+++ lib/pages/chat/channel/publish_channel_list_widget.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:colla_chat/constant/base.dart';
+import 'package:colla_chat/crypto/util.dart';
 import 'package:colla_chat/datastore/datastore.dart';
 import 'package:colla_chat/entity/chat/chat_message.dart';
 import 'package:colla_chat/l10n/localization.dart';
@@ -9,8 +10,12 @@ import 'package:colla_chat/pages/chat/channel/publish_channel_edit_widget.dart';
 import 'package:colla_chat/plugin/logger.dart';
 import 'package:colla_chat/provider/index_widget_provider.dart';
 import 'package:colla_chat/service/chat/chat_message.dart';
+import 'package:colla_chat/service/chat/message_attachment.dart';
 import 'package:colla_chat/tool/date_util.dart';
+import 'package:colla_chat/tool/dialog_util.dart';
+import 'package:colla_chat/tool/document_util.dart';
 import 'package:colla_chat/tool/image_util.dart';
+import 'package:colla_chat/tool/json_util.dart';
 import 'package:colla_chat/widgets/common/app_bar_view.dart';
 import 'package:colla_chat/widgets/common/common_widget.dart';
 import 'package:colla_chat/widgets/common/widget_mixin.dart';
@@ -112,6 +117,32 @@ class _PublishChannelListWidgetState extends State<PublishChannelListWidget>
             curve: Curves.easeIn));
   }
 
+  ///将编辑的内容正式发布，统一采用html格式保存和发送，原先保存的草案要转换格式，更新状态
+  _publish(ChatMessage chatMessage) async {
+    String? mimeType = chatMessage.mimeType;
+    if (mimeType == ChatMessageMimeType.json.name) {
+      var bytes = await messageAttachmentService.findContent(
+          chatMessage.messageId!, chatMessage.title!);
+      if (bytes != null) {
+        String json = CryptoUtil.utf8ToString(bytes);
+        var deltaJson = JsonUtil.toJson(json);
+        var html = DocumentUtil.jsonToHtml(deltaJson);
+        chatMessage.mimeType = ChatMessageMimeType.html.name;
+        chatMessage.content = chatMessageService.processContent(html);
+      }
+      chatMessage.status = MessageStatus.published.name;
+      chatMessage.sendTime = DateUtil.currentDate();
+      await chatMessageService.store(chatMessage);
+    } else if (mimeType == ChatMessageMimeType.html.name) {
+      await myChannelChatMessageController.publish(chatMessage.messageId!);
+    }
+
+    if (mounted) {
+      DialogUtil.info(context,
+          content: AppLocalizations.t('Publish document successfully'));
+    }
+  }
+
   ///展示每一条消息
   Widget _buildChannelChatMessageItem(BuildContext context, int index) {
     List<ChatMessage> chatMessages = myChannelChatMessageController.data;
@@ -154,22 +185,42 @@ class _PublishChannelListWidgetState extends State<PublishChannelListWidget>
     );
 
     ActionPane actionPane = ActionPane(
-      extentRatio: 0.3,
+      extentRatio: 0.6,
       motion: const DrawerMotion(),
       children: <SlidableAction>[
         SlidableAction(
           padding: const EdgeInsets.symmetric(vertical: 0.0, horizontal: 0),
           onPressed: (context) async {
-            await chatMessageService
-                .remove(where: 'id=?', whereArgs: [chatMessage.id!]);
-            myChannelChatMessageController.delete(index: index);
+            bool? confirm = await DialogUtil.confirm(context,
+                content: 'Do you confirm publish?');
+            if (confirm != null && confirm) {
+              _publish(chatMessage);
+              setState(() {});
+            }
           },
           backgroundColor: Colors.blue,
+          foregroundColor: Colors.white,
+          icon: Icons.publish,
+          label: AppLocalizations.t('Publish'),
+          borderRadius: BorderRadius.circular(0),
+        ),
+        SlidableAction(
+          padding: const EdgeInsets.symmetric(vertical: 0.0, horizontal: 0),
+          onPressed: (context) async {
+            bool? confirm = await DialogUtil.confirm(context,
+                content: 'Do you confirm delete?');
+            if (confirm != null && confirm) {
+              await chatMessageService
+                  .remove(where: 'id=?', whereArgs: [chatMessage.id!]);
+              myChannelChatMessageController.delete(index: index);
+            }
+          },
+          backgroundColor: Colors.amber,
           foregroundColor: Colors.white,
           icon: Icons.delete_outline,
           label: AppLocalizations.t('Delete'),
           borderRadius: BorderRadius.circular(0),
-        )
+        ),
       ],
     );
